+++ packages/core/src/components/_parent.ts
@@ -1,4 +1,4 @@
-import ErrorCompotes from '../../utils/error'
+import ErrorCompotes from '../utils/error'
 
 export interface ParentOptions {
   /**
